pyproject.toml
[tool.poetry]
name = "nonebot-plugin-test"
version = "0.1.0"
description = "Test frontend for nonebot v2+"
authors = ["yanyongyu <yanyongyu_1@126.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://v2.nonebot.dev/"
repository = "https://github.com/nonebot/plugin-test"
keywords = ["nonebot", "qq", "qqbot", "mirai", "cqhttp"]
include = ["nonebot_plugin_test/dist/**/*"]

[tool.poetry.dependencies]
python = "^3.7"
aiofiles = "^0.5.0"
python-socketio = "^4.6.0"
nonebot2 = "^2.0.0-alpha.6"

[tool.poetry.dev-dependencies]
yapf = "^0.30.0"

[[tool.poetry.source]]
name = "aliyun"
url = "https://mirrors.aliyun.com/pypi/simple/"
default = true

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
